+++ bin/check-mysql-status.rb
@@ -115,7 +115,11 @@ class CheckMySQLStatus < Sensu::Plugin::Check::CLI
       section = ini[config[:ini_section]]
       db_user = section['user']
       db_pass = section['password']
-      db_socket = section['socket']
+      db_socket = if config[:socket]
+                    config[:socket]
+                  else
+                    section['socket']
+                  end
     else
       db_user = config[:user]
       db_pass = config[:password]